### 2021-03-16 #435 @ekothe
gendercoder: Recodes Sex/Gender Descriptions Into A Standard Set
Submitting Author: Emily Kothe (@ekothe)
Other Authors: Jennifer Beaudry (@jlbeaudry), Felix Singleton Thorn (@fsingletonthorn), Rhydwyn McGuire (@rhydwyn), Mathew Ling (@Lingtax), Nicholas Tierney (@njtierney)
Repository: https://github.com/ropenscilabs/gendercoder
Version submitted: 0.0.0.9000.

Version accepted: TBD

---

```
Package: gendercoder
Title: Recodes Sex/Gender Descriptions Into A Standard Set
Version: 0.0.0.9000
  person("Jennifer", "Beaudry", role = "aut",
    email = "jbeaudry@swin.edu.au",
    comment = c(ORCID = "https://orcid.org/0000-0003-1596-6708")),
  person("Emily", "Kothe", role = c("aut", "cre"),
    email = "emily.kothe@deakin.edu.au",
    comment = c(ORCID = "https://orcid.org/0000-0003-1210-0554")),
  person("Felix", "Singleton Thorn", role = "aut",
    email = "fsingletonthorn@gmail.com",
    comment = c(ORCID = "https://orcid.org/0000-0002-0237-6146")),
  person("Rhydwyn", "McGuire", role = "aut",
    email = "rhydwyn@rhydwyn.net"),
  person("Nicholas", "Tierney",
         role = c("aut"),
         email = "nicholas.tierney@gmail.com",
         comment = c(ORCID = "https://orcid.org/0000-0003-1460-8722")),
  person("Mathew", "Ling",
         role = c("aut"),
         email = "mathewtyling@gmail.com",
         comment = c(ORCID = "https://orcid.org/0000-0002-0940-2538"))
         )
Description: Provides functions and dictionaries for recoding of freetext gender responses into more consistent categories.
Maintainer: Emily Kothe <emily.kothe@deakin.edu.au>
    purrr
    covr

```

The goal of gendercoder is to allow simple re-coding of free-text gender responses. This is an important data munging task for researchers who collect gender in this format.

Best practice recommendations for the collection of gender information from human participants includes collection of some or all responses in free-text. This allows participants to provide gender responses when their gender is not captured by researcher provided categories (e.g. male, female, non-binary) or when gender is collected without researcher provided categories (e.g. “What is your gender”). gendercoder contains two in-built output dictionaries, a default broad dictionary which corrects spelling and standardises terms while maintaining the diversity of responses and a narrow dictionary which contains only three gender categories, “male”, “female”, and “sex and gender diverse”.

To our knowledge this is the only package of its kind.

N/A

N/A
