### 2021-03-17 576031a
Fix for issue #26
The purrr/tolower at line 60 only applied to values, not to "names".
