+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
   rationale: rationale.html
-last_built: 2021-03-16T21:31Z
+last_built: 2021-03-17T23:27Z
 
