### 2021-03-29 e2bd9a9
Adds the gendercoder_app() function which initiates a local instance of the gendercoder shiny app
