### 2021-04-05 #435 @noamross
Hello @ekothe and team, sorry to take so long to get back to you on this.  The editors have gone a couple of rounds discussing this, and we decided that it falls out of scope for us.  (Ultimately it comes down to whether this is "data munging of scientific formats").  Lots of packages that we like, and come out of the unconferences end up out of scope!  We hope you continue with this project.
