+++ README.Rmd
@@ -67,6 +67,16 @@ tibble(gender = c("male", "MALE", "mle", "I am male", "femail", "female", "enby"
   
 ```
 
+The package now also contains a shiny app for use of the `gender_recode()` function by GUI. The app takes data in `.dta`, `.csv`, and `.sav` formats.
+
+To use the app, call the `gendercoder_app()` function.
+
+```{r, eval=FALSE}
+gendercoder::gendercoder_app()
+```
+
+<img src="man/figures/UI.png">
+
 
 ## Contributing to this package
 
