+++ man/narrow.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gendercodeR.R
+% Please edit documentation in R/gendercoder.R
 \docType{data}
 \name{narrow}
 \alias{narrow}
