### 2021-04-08 101af9f
Make compliant for CRAN submission.
