+++ cran-comments.md
@@ -1,0 +1,12 @@
+## Test environments
+* macOS X Catalina (gh-actions), R 4.0.5
+* ubuntu 20.04 (gh-actions), R 4.0.5
+* ubuntu 20.04 (gh-actions), dev 2021-04-05 r8014
+* windows server 19 (gh-actions), R 4.0.5
+
+## R CMD check results
+Note: This is a new submission.
+
+## Downstream dependencies
+
+There are currently no downstream dependencies for this package
