### 2021-04-08 #435 @ekothe
Hi @noamross I wonder if this discussion led to any clarity about what constitutes an in scope "scientific format" that could be shared somewhere to save similar submissions in the future? Reading the [package categories guidance](https://devguide.ropensci.org/policies.html#package-categories) it seems to us that insofar as gendercoder processes data from "unstructured sources such as text", and is used "...for handling data in specific scientific formats" it complies with the scope for data munging packages.

While the scope of the package is quite narrow, it has somewhat parallel functionality to [coder ](https://docs.ropensci.org/coder/), and we would speculate that by contrast the use cases of gendercoder are much wider and central to ropensci's mission of creating software tools that lower barriers to working with scientific data and that promotes a welcoming and diverse community given that most studies of human participants would collection of gender as a key demographic variable.

In this context, I hope you can understand our confusion and hope you can provide public guidance on why a package like this is out of scope to save avoid other submissions encountering similar issues in the future?
