### 2021-04-15 #435 @noamross
Hi @ekothe. This package was certainly an edge case, and it had come down to a matter of specificity vs. generality. Our scope for data extraction and munging packages aims to be those that apply to workflows that come out of specific formats vs. general sources. The reason for this is that we think its easier define a "standard" way to handle a specific format or structure, while preferences for workflows tend to be very subjective.  Our interpretation in this case was that unstructured survey responses were a general type of data rather than specific format/source.  I think the best improvement we could make to the documentation is to list more illustrative examples, and I will make a note to do so.

However, the editors took another link and found your analogy to the **coder** package compelling enough to reconsider our evaluation, so we have reversed our decision.  Our scope definitions are always evolving and somewhat path-dependent.  Thanks for making a clear argument. And as I said, we've always been very supportive of this project and the goals behind it!
