### 2021-04-15 #435 @ekothe
Thank you @noamross, I appreciate the update and look forward to working with @ldecicco-USGS as the editor 😄
