### 2021-04-28 #435 @ldecicco-USGS
Sorry for all the bot messages! I ran the checks/goodpractices and everything looks good:
```
Preparing: description
Preparing: lintr
Preparing: namespace
Preparing: rcmdcheck
── GP gendercoder ───────────────────────────────────────────────────────────────

It is good practice to

    inst\app\server.R:28:1
    inst\app\server.R:29:1
    inst\app\server.R:48:1
    inst\app\ui.R:43:1
    inst\app\ui.R:44:1
    ... and 7 more lines

    vapply() instead.

    R\gendercoder_app.R:16:16
```
I wouldn't worry too much about the long lines, especially since they seem to all be in the shiny app.  Let's see if the next "bot" works:
