### 2021-04-28 #435 @elinw
Looking forward to digging into this, having recently asked an open ended gender question on a survey.  Putting this reference here too since it highlights the important issues http://www.stat.columbia.edu/~gelman/research/unpublished/Using_gender_in_surveys.pdf.
