### 2021-04-28 #435 @elinw
Two initial questions.

Why use purr rather than a hash table?  See https://blog.dominodatalab.com/a-quick-benchmark-of-hashtable-implementations-in-r/  for a quick summary.

Should you name your dictionaries in a way that identifies them as English?  Using this same pattern people will be able to make (and potentially contribute if you want to accept them) dictionaries in other languages.

For my recent survey I went from
 - original: 1499 unique values and 1499 NA
 -  to narrow 4 unique values and 1541 NAs
 - to broad 44 unique and 1499 NA

(It also matched perfectly with my recoding.)
