### 2021-05-04 #435 @elinw
One way to do a hash would be to use the data.table hash form qdapTools.

This didn't give me totally the results I expected because it  returned 1541 missing (as with narrow originally), but exploring options and maybe some pre or post processing would probably fix.

# Get the data and create a data frame instead
value_broad <- sapply(broad, "[[", 1)
value_narrow <- sapply(narrow, "[[", 1)
key_broad <- names(value_broad)
key_narrow <- names(value_narrow)
value_broad <- unname(value_broad)
value_narrow <- unname(value_narrow)
broad_df <- data.frame("key" = key_broad, "value" = value_broad)

# my data set is called students and the variable is called gender.
broad_hash <- qdapTools::hash(broad_df)
students$new_gender <- qdapTools::hash_look(students$gender, broad_hash)
