### 2021-05-04 #435 @Lingtax
I have a partially functioning version with the base::environment hash method in a branch on my fork now, I just need to work out how to make the fill argument work again in that format. Assuming that method is the fastest as per the blog you forwarded, @elinw.

The only other hiccup will be supporting users in adding their own dictionaries or augmenting the inbuilts, as I found building the hash table was slightly more frustrating than I expected.

Get Outlook for iOS<https://aka.ms/o0ukef>
