### 2021-05-04 #435 @juliasilge
> Why use purr rather than a hash table?

I think a really good reason for this is maintainability, ease of use, ease of user contributions (more important in this package than normal), and exactly the frustrations that @Lingtax mentions. I wouldn't necessarily jump to something like a hash table, which is a new idea to many R users (trained not as CS folks), compared to purrr, which is like `lapply()` but type safe. How big are people's surveys expected to be? What kind of performance gains are we talking about?
