### 2021-05-04 #435 @Lingtax
Another point is that IIRC, the core functionality isn't actually purrr based, I think Nick replaced that with a list comprehension method. purrr is only in there to make the dictionaries consistent to support the matching (ie. making them all lower case).
