### 2021-05-12 #435 @elinw
I've been looking into this issue exactly and I'll try to share some code in the next couple of days.  I think for users you would provide a function that takes a data frame and turns it into a hash.
