### 2021-05-12 #435 @Lingtax
I think the original format was closer to if not a named vector, @ekothe might recall why the change was made. Agree that the list is slightly less intuitive than ideal.
