### 2021-05-13 #435 @elinw
Here's some sample code that is not in the form of functions, that illustrate what you could do staying in base R. It's not as efficient as it could.
https://gist.github.com/elinw/fc73a5555ea6706a92b437715036093e

In my example I have a variable students$gender and I'm recoding it into students$gender2.
