### 2021-05-13 #435 @Lingtax
Thanks Elin. That's similar to what I had, though the fill is closer than mine got. I'm not getting 100% yet, at present the fill in that gist fills using the first value of gender or fails (not a multiple of replacement length).

There's still going to be friction for users augmenting the in-built dictionaries with this pipeline though. At present users can just concatenate their custom options to the list in the call, e.g.
`recode_gender(gender, dictionary = c(broad, list("amel" = "male"))`

We could move to a hash if that was the right option, but then we would need to make a new hash table after the function is called or update the old one (I have no idea, I've never deal with these before now), assuming the hash table comes in the package as an existing data object.
