### 2021-05-13 #435 @elinw
I actually updated slightly based on reading about performance.
You could store the hashed dictionaries like you do now.   I was thinking about the concatenation issue and I don't think it is a deal breaker, you could let  people add to the dictionary with  a function  that would automate broad[["amel"]] <- "male" .   e.g. `add_terms(key, value, dictionary)`.

I really think this package is going to be useful.  In my same survey I have an open-ended race/ethnicity question so this is a very general problem.   Not right now, but that's something else to think about longer term.
