### 2021-05-17 #435 @elinw
So having dived into this some more, I actually profiled using a named vector and that was faster. I'm sure it depends on the size of the dictionary (if it had a million entries it would be slower) but maybe at a scale for this kind of action that would make sense.
