### 2021-05-18 #435 @ekothe
The question about performance gains of different methods is really interesting to me. However, I do think some context around expected size of both the data sets and dictionaries might help to place the expected gains from any kind of refactoring.

>... How big are people's surveys expected to be? What kind of performance gains are we talking about? I worked with surveys with 100k+ free text responses and using something like purrr would not be a bottleneck.

It is hard to get clear numbers on this, but it might be useful to consider something like the UK Census to get a reasonable upper limit on dataset size. The UK Census has a free text gender response, but it is only completed by people who indicate that their sex assigned at birth is different from their gender identity (afaik it's the largest English language census with this kind of question). Stonewall UK estimates that about 1% of the UK population is trans, given the last UK Census was completed by approximately 56 million people, we could estimate that 560,000 people could provide a free-text response to this item (_note that this is a dubious calculation because the question asks about legal gender and so some trans people will not provide a free text response and there has been some negative publicity that might add to malicious responses - however it provides a starting point for scoping the issue_). I think in general, with current approaches at least, most people who are collecting gender identity from a very large number of people ask the question in such a way that most participants can provide a response from a forced choice (e.g. male/female) and only a subset of participants provide some free-text information. In data sets where all gender is recorded in free-text I would expect that the (small) number of responses would be completely undetectable (e.g. in psychology I'd expect to see datasets with 100-1000 participants).

>So having dived into this some more, I actually profiled using a named vector and that was faster. I'm sure it depends on the size of the dictionary (if it had a million entries it would be slower) but maybe at a scale for this kind of action that would make sense.

In terms of the size of the dictionary, the most recent "Gender Census" (see gendercensus.com) collected data from 44,583 people who reported that their gender identity is not fully captured by the gender binary. From this there were 7,994 unique textbox entries for gender identity. With many of these being responses that would never be in a dictionary like `gendercoder::` (e.g. "just [name]").
