### 2021-05-22 #435 @juliasilge
I recommend adding a full descriptive title to the vignette. Also I would edit to always use "built-in" or "inbuilt" but not switch back and forth.

- [ ] **Examples** (that run successfully locally) for all exported functions

I think there is a mistake in the example for the app; it says `scivrs_app()`.

I recommend adding a second, short vignette on how, when, why to add to an existing dictionary or add a new dictionary, based on the content in `data/howtoaddtodictionary.R`. You can link to that from the README's contributing section.

I know there has been some discussion of performance previously on this thread; my opinion is that `purrr::map(dictionary, tolower)` and `purrr::flatten_chr(recoded_list)` are very reasonable in this context (purrr is pretty darn lightweight) and that the use of lists for matching is great. 👍 My opinion is that this isn't a package whose applications make it a sensible place to invest more effort in performance improvements beyond what is already here.

Can you add the [coverage](https://codecov.io/github/ropenscilabs/gendercoder/) badge to your README? You don't have any tests for the app, which isn't great; the [rOpenSci package guide recommends shinytest](https://devguide.ropensci.org/building.html#testing).

Something is up with the [pkgdown](https://ropenscilabs.github.io/gendercoder/) site, as it isn't up-to-date with the package (no info on the app).

---

I'd like to congratulate the package authors on a thoughtful handling of this important topic and for providing a tool that will help more survey creators and analysts ask free-text gender questions. This package is in excellent shape, and I think the choice of the main function API is good (input a character vector, return a character vector).

Did you consider any other names for the `fill` argument? It means something somewhat different in tidyr, which may be confusing to users.

My first major feedback is that this package uses the terms "male" and "female" for gender, but in my experience most practitioners are moving away from these terms for gender, instead of using them for sex. If you think about how people are more deeply understanding [sex vs. gender](https://en.wikipedia.org/wiki/Sex_and_gender_distinction), this package seems to be more about gender than sex (this is typically what survey questions are getting at) and better terms for that would be "man" and "woman". I strongly recommend changing this throughout the documentation and the data dictionaries.

The second major feedback is that I find the terms "broad" and "narrow" so confusing because to me, I would have flipped them, with the broad dictionary bucketing responses into "broadly" into categories all together and the narrow dictionary keeping lots of detailed categories. Is this just me? Did you all talk about any other names for the two options that might be more "broadly" 😩 understood?

Another note is that the discussions of context in the vignette and README are great, but I would also add an acknowledgement that our understanding is changing with _time_ and that is expected and OK, etc. You can commit to changing these mappings as our understanding grows/changes.

Again, amazing job on creating this tool and making it available to the data science community. 🙌 I wish I would have had it when I was working on the Stack Overflow survey!
