### 2021-05-22 7533e66
Edits/typos/consistency in README, vignette, docs
