+++ vignettes/rationale.Rmd
@@ -10,7 +10,7 @@ vignette: >
 
 # gendercoder
 
-The goal of gendercoder is to allow simple re-coding of free-text gender responses.
+The goal of gendercoder is to allow simple recoding of free-text gender responses.
 
 ## Installation
 
@@ -162,8 +162,8 @@ str(custom_dictionary)
 
 Custom dictionaries can be used in place of an inbuilt dictionary or can 
 supplement the inbuilt dictionary by providing a vector of lists to the 
-dictionary argument. Where the lists contain duplicated elements the last 
-version of the duplicated value will be used for re-coding. This allows you to 
+dictionary argument. Where the lists contain duplicated elements, the last 
+version of the duplicated value will be used for recoding. This allows you to 
 use the inbuilt dictionary but change the coding of one or more responses from 
 that dictionary. Here the addition of Spanish terms allows for recoding of 11 
 previously uncoded responses.
@@ -190,7 +190,7 @@ sample %>%
 
 ### fill
 
-The `fill` argument is used to determine the handling the re-coding of values
+The `fill` argument is used to determine the handling for recoding of values
 not contained in the dictionary. By default, unmatched values are coded as NA. 
 `fill = TRUE` will fill unmatched responses with the participant provided 
 response. 
@@ -207,33 +207,32 @@ sample %>%
   knitr::kable(caption = "Summary of gender categories after use of the combined dictionary and `fill = TRUE`")
 ```
 
-# A disclaimer on gender
+# A disclaimer on handling gender responses
 
 This package attempts to remove typographical errors from free text gender data.
-The defaults that we used are specific to our context and your data may be
-different. 
+The defaults that we used are specific to our context and your data or context 
+may be different. 
 
-We offer two built-in dictionaries, board and narrow. Both are opinionated
-about how gender descriptors collapse into categories. 
+We offer two built-in dictionaries, broad and narrow. Both are necessarily 
+opinionated about how gender descriptors collapse into categories. 
 
-However, as these are culturally specific they may not be suitable for your data. 
-In particularly the narrow setting makes opinionated choices about some responses 
+However, as these are culturally specific, they may not be suitable for your data. 
+In particular the narrow option makes opinionated choices about some responses 
 that we want to acknowledge are potentially problematic. In particular,
 
-- In 'narrow' coding intersex responses are recoded as 'sex and gender
+- With narrow coding, intersex responses are recoded as 'sex and gender
 diverse'
-- In 'narrow' responses where people indicate they are trans and
-indicate their identified gender are recoded as the identified gender
-(e.g. 'Male to Female' is recoded as Female). We wish to acknowledge
-that this may not reflect how some individuals would classify
-themselves when given these categories and in some contexts may make
-systematic errors. The broad coding dictionary attempts to avoid these
-issues as much as possible - however users can provide a custom
-dictionary to add to or overwrite our coding decisions if they feel
-this is more appropriate. We welcome people to update the inbuilt
+- With narrow coding, responses of people who indicate they are trans plus 
+their gender are recoded as the identified gender (e.g. 'Male to Female' is 
+recoded as Female). We wish to acknowledge that this may not reflect how some 
+individuals would classify themselves when given these categories; in some 
+contexts this recoding may make systematic errors. The broad coding dictionary 
+attempts to avoid these issues as much as possible; however users can provide a 
+custom dictionary to add to or overwrite our coding decisions if they feel
+this is more appropriate. We welcome contributions to update the inbuilt
 dictionary where desired responses are missing.
-- In both dictionaries we assume that typographical features such as spacing
-are not relevant to re-coding of the gender response (e.g. we assume that
[…]
