### 2021-05-24 c84e7a6
correct documentation and add very minimal testing to app.
