### 2021-05-24 fb18290
update app tests for comprehensiveness, integrates with testthat
