### 2021-05-24 #435 @Lingtax
Thanks Julia. Apologies as I introduced the app as a half-baked idea after review had already started hence the bugs, poor documentation, and lack of tests. I've added some commits that hopefully will address those.

We'll address the documentation and vignette more extensively and have a discussion about how to proceed with the terminology. I agree re: "man / woman" and "fill", though unsure what to do on the latter front.  `retain_unmatched` is more explicit, but it's quite clunky. Regarding broad and narrow, I personally find the current terminology to be intuitive, but I don't know idiographic it is, so good to note and we might do some polling.

Thank you again for your time and feedback, it's greatly appreciated.
