+++ DESCRIPTION
@@ -33,6 +33,7 @@ Imports:
     purrr
 Suggests: 
     testthat (>= 2.1.0),
+    shinytest,
     knitr,
     rmarkdown,
     dplyr,
