### 2021-05-24 #435 @ldecicco-USGS
Thanks @juliasilge for the thorough review! @elinw , you've provided some feedback - I just wanted to check that you are also planning to submit a final review (preferably using the template found [here](https://devguide.ropensci.org/reviewtemplate.html)).
