+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
   rationale: rationale.html
-last_built: 2021-04-15T23:43Z
+last_built: 2021-05-25T04:08Z
 
