### 2021-06-10 11f48e5
Update run-tests.yaml
Add lines to install phantomjs dependency
