### 2021-06-10 2d10761
converts outputs from male and female to man or boy and woman or girl
