### 2021-06-10 506f936
Adds _en suffixed dictionaries and updates functions and documentation to make default.
