+++ data/howtoaddtodictionary.R
@@ -6,6 +6,9 @@
 #  males = "male",
 #  females = "female",
 #  boi = "sex and gender diverse")
+#
+# dictionary_df <- read.csv("data.csv")
+# dictionary_from_df <- as.list(tibble::deframe(dictionary_df))
 
 #new <- append(narrow, custom_dictionary)
 #narrow <- new
