+++ man/broad_en.Rd
@@ -1,12 +1,9 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/gendercodeR.R
 \docType{data}
-\name{broad}
-\alias{broad}
-\title{broad}
+\name{broad_en}
+\alias{broad_en}
+\title{broad_en}
 \description{
-This dictionary is deprecated, please use broad_en instead.
-}
-\details{
 A broad English dictionary for the recode_gender function in list form
 }
