### 2021-06-10 ebbbccc
Update check-standard.yaml
Add phantomJS yaml
