### 2021-06-10 #435 @elinw
Here you go:

I really like the concept of this package and think it will get a lot of use as the asking of open-ended gender items on surveys becomes increasingly common.  This is part of a larger category of such questions. For example I recently asked about race/ethnicity in this format. I also asked the incoming cohort of first year students what major they were thinking of and that also needed coding. Further, many surveys include an “other ____” category which is often difficult to deal with.  I’d imagine that this package would be extremely helpful.

Building on this comment, in reality, someone could add any dictionaries they wanted to this.  That is, there is nothing inherently about gender. I think you should consider whether you rename it to something more generic and then simply say you are including the gender dictionaries as examples.   I think that small change will have a huge impact on use.

The other point I would make about the dictionaries you have is that they are in English. Obviously this is not something that only applies in Anglophone settings.  I think that you should change the names of the dictionaries to reflect the language in some consistent pattern.  I’d suggest using the ISO IDs so you can differentiate EN-AU, EN-UK and EN-US, or more importantly different versions of Spanish and Chinese.

Shiny App

I attempted to pull in a large  csv file which meant I couldn’t see the gender column. I’m sure there is a solution to that (maybe just display the gender column once it is selected?).

I was also surprised that it didn’t support Rda or Rds file types.

Documentation of the use of the Shiny appwould b helpful.

Coding

As I mentioned I do not think a list is the best way to store the dictionaries for a number of reasons. First, based on my testing that using a named vector instead gets you substantial performance improvement.  I didn’t see any major performance benefit of using a hash table, so I think the ease of use of a named vector makes it the way to go, though in a much larger file there would be a payoff.  You could rewrite the function like this

```
gender_named_vector <- function(x, nv, fill){
	  recoded<- nv[trimws(tolower(x))]
	  if (fill == TRUE){
	    recoded <- ifelse(is.na(recoded), x, recoded)
	  }
	  unname( recoded )
	}
```

If you have a vector, a lot of other things become simpler. For example you just can use `tolower()` to make all of the characters lower case.

One of the nice things about this package is that the code is very straightforward an cleanly written.

You have examples and documentation for the internal function ` which_is_na()`.  This means you’ll get a manual page that you don’t want. Also I think that CRAN  will flag that.  Also in the examples you should namespace the external functions. CRAN would also probably also like to start the function descriptions so that it has an upper case first letter.

I will definitely be recommending this package.
