### 2021-06-10 #435 @elinw
Also I agree with @juliasilge  about broad and narrow, I thought they would have the opposite meanings. I also agree that you need to think of your dictionaries as living because these ideas are changing and that's how it is both with language and social constructions.
