### 2021-06-10 86e5548
Moving installation of rcmdcheck
