+++ .github/workflows/check-standard.yaml
@@ -65,8 +65,8 @@ jobs:
 
       - name: Install dependencies
         run: |
-          remotes::install_deps(dependencies = TRUE)
           remotes::install_cran("rcmdcheck")
+          remotes::install_deps(dependencies = TRUE)
         shell: Rscript {0}
 
       - name: Find PhantomJS path
