+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: gendercoder
-Title: Recodes Sex/Gender Descriptions Into A Standard Set
+Title: Recodes Sex/Gender Descriptions into a Standard Set
 Version: 0.0.0.9000
 Authors@R: c(
   person("Jennifer", "Beaudry", role = "aut",
