### 2021-06-10 #435 @ldecicco-USGS
Thanks for the review @elinw 🎉 To add another option to the efficiency mix, was curious if the `ifelse` was the best way to go. Something like:
```r
library(dplyr)

# here's a dictionary with the dplyr "starwars" data frame:
df_dictionary <- starwars %>%
  select(homeworld) %>%
  distinct() %>%
  na.omit() %>%
  mutate(homeworld = homeworld,
         new_name = c(paste0("first_", letters),
                      paste0("second_", letters[1:22])))

#Let's take out 3 worlds to make sure the "fill" works
df_dictionary <- df_dictionary[-1:-3,]

repeat_rows <- 1000000
starwars_big <- data.frame(homeworld = rep(starwars$homeworld,
                                           repeat_rows),
                           species = rep(starwars$species, repeat_rows))

nv <- setNames(df_dictionary$new_name, df_dictionary$homeworld)

gender_named_vector_a <- function(x, nv, fill){
  recoded <- nv[x]
  if (fill == TRUE){
    recoded <- ifelse(is.na(recoded), x, recoded)
  }
  unname( recoded )
}

gender_named_vector_b <- function(x, nv, fill){
  recoded<- nv[x]
  if (fill == TRUE){
    recoded[is.na(recoded)] <-  x[is.na(recoded)]
  }
  unname( recoded )
}

system.time({
  x <- gender_named_vector_a(starwars_big$homeworld, nv,
                              fill = TRUE)
})
   user  system elapsed
  22.22    1.36   23.58
system.time({
  y <- gender_named_vector_b(starwars_big$homeworld, nv,
                              fill = TRUE)
})
   user  system elapsed
   3.91    0.83    4.73
```

Otherwise, I did a few tests comparing a named vector, to `dplyr`'s left_join and joins using `data.table`. All were pretty darn similar to the named vector. So, if it would make more sense to keep the data in a data frame, something like this could work:
```r
dp_join_sm <- function(df, df_dict, key, fill){

  df_recode <- df %>%
    left_join(df_dict, by = key, na_matches = "never")

  if (fill == TRUE){
    df_recode$new_name[is.na(df_recode$new_name)] <-  df_recode[[key]][is.na(df_recode$new_name)]
  }

  df_recode <- df_recode[,-which(names(df_recode) == key)]
  names(df_recode)[names(df_recode) == "new_name"] <- key

  return(df_recode)
}

system.time({
   z <- dp_join(starwars_big, df_dictionary,
                   key = "homeworld", fill = TRUE)
 })
   user  system elapsed
   6.56    0.86    7.42

```

Anyway, just some morning futzing... I know recoding text has always been a bit of a bottleneck for me, so it's always good to re-check performance from time to time.

The ball's back in @ekothe and team's court for a response to reviewer.
