### 2021-06-12 #435 @elinw
The fill i think could also be handled in a different way, but let me write up something.  A few clarifications.

- by size, I was really meaning the number of rows being recoded not the size of the dictionary. That's obviously a different issue.
- I don't think adding values to any version is complicated. I've learned so much about this issue by doing the review, but even with the hash table all you are doing is adding dict[[key]] <-value which could easily be made into a function that takes multiple pairs,  (I feel I should write a blog about this whole hash versus named vector issue).
- I just installed a new package that also has a kind of recoding (questionr)  and they have RStudio add ins ... I wonder if making an add in would be better than having the shiny app..
