### 2021-06-12 9250b80
remove docs for which_is_na. Issue #44
