+++ R/recode_gender.R
@@ -1,18 +1,3 @@
-#' finds terms without a dictionary match
-#'
-#' Internal function to identify responses that are not
-#' matched in the dictionary.
-#'
-#' @param x a list that contains responses and corresponding dictionary
-#' matches #' responses without a match are stored with the NA or NULL
-#' as the list reference.
-#'
-#' @return a vector of positions in input list where name is missing
-#'
-#' @examples
-#' x <- list("a"=1,"b" =2, "c" =3)
-#' names(x)[c(1,3)] <-  NA
-#' gendercoder:::which_is_na(x)
 which_is_na <- function(x) {
   which(is.na(names(x)))
 }
