### 2021-06-12 02225cc
replace `fill` argument with `retain_unmatched` argument. Update documentation. Closes #47
