+++ README.Rmd
@@ -51,7 +51,7 @@ The core function, `gender_recode()`, takes 3 arguments,
 
 - `dictionary` the preferred dictionary, and
 
-- `fill` a logical indicating whether original values should be carried over if 
+- `retain_unmatched` a logical indicating whether original values should be carried over if 
 there is no match. 
 
 Basic usage is demonstrated below. 
@@ -60,8 +60,8 @@ Basic usage is demonstrated below.
 library(gendercoder)
 
 tibble(gender = c("male", "MALE", "mle", "I am male", "femail", "female", "enby")) %>% 
-  mutate(broad_gender  = recode_gender(gender, dictionary = broad, fill = TRUE),
-         narrow_gender = recode_gender(gender, dictionary = narrow, fill = FALSE)
+  mutate(broad_gender  = recode_gender(gender, dictionary = broad, retain_unmatched = TRUE),
+         narrow_gender = recode_gender(gender, dictionary = narrow, retain_unmatched = FALSE)
   )
   
 ```
