### 2021-06-12 42e150f
Converts the core function to operate on dictionaries in named vectors (Cheers Elin and Laura). Updates documentation to match.
