+++ man/gendercoder.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gendercoder.R
+% Please edit documentation in R/gendercodeR.R
 \docType{package}
 \name{gendercoder}
 \alias{gendercoder}
