+++ man/narrow.Rd
@@ -1,9 +1,9 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gendercoder.R
+% Please edit documentation in R/gendercodeR.R
 \docType{data}
 \name{narrow}
 \alias{narrow}
 \title{narrow}
 \description{
-A narrow dictionary for the recode_gender function in list form
+A narrow dictionary for the recode_gender function in named vector form
 }
