+++ man/broad.Rd
@@ -1,9 +1,9 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gendercoder.R
+% Please edit documentation in R/gendercodeR.R
 \docType{data}
 \name{broad}
 \alias{broad}
 \title{broad}
 \description{
-A broad dictionary for the recode_gender function in list form
+A broad dictionary for the recode_gender function in named vector form
 }
