### 2021-06-12 b153f06
Updates Shiny app so it preferentially displays the selected variable and the recoded variable. #46
