### 2021-06-12 cb1ad84
add support for rds in shiny app.
