### 2021-06-12 1dcebb8
make function documentation independent of dplyr #44
