### 2021-06-12 04cc695
remove pipes from all core functions.
