### 2021-06-24 432b45a
Adding a short vignette on adding to the dictionary.
