### 2021-06-27 #435 @elinw
One reason it was confusing to me is that there is also a dimension of
"restrictive" and "something opposite"  meaning that I wondered if narrow
would have left more of the responses unclassified compared to broad.

On Fri, Jun 11, 2021 at 10:53 PM Emily Kothe ***@***.***>
wrote:

> Thank you @elinw <https://github.com/elinw> and @juliasilge
> <https://github.com/juliasilge> for your reviews (and to @ldecicco-USGS
> <https://github.com/ldecicco-USGS> for you contribution as well).
>
> We will be working on this over the next couple of weeks.
>
> There are a number of interesting issues raised, in particular, I notice
> that @elinw <https://github.com/elinw> and @juliasilge
> <https://github.com/juliasilge> have both flagged that "broad" and
> "narrow" had the opposite meaning for them than it does for the gendercoder
> team. I wonder if this is reflective of Australian English or simply that
> when creating the dictionaries we were thinking in terms of the number of
> categories that inputs would be coded into (such that broad = a large
> number of categories) whereas broad/narrow would have the opposite meaning
> if you conceptualise it in terms of the exactness of the matching. Because
> there seems to be such conflicting interpretation of that terminology it is
> obvious that we can't simply resolve it by reversing the terms, so we need
> to put some thought into a naming convention that is more interpretable to
> a wider range of people. I would note that we originally called narrow
> "three categories" but realised that that would be problematic if we ever
> needed to rework it so that there were more than three output categories of
> gender in the dictionary. We might need to elicit suggestions from a
> broader community of people to see if we can come up with something that
> reflects the intended usage of each dictionary more clearly.
>
