+++ inst/app/tests/shinytest/app_test.R
@@ -3,6 +3,6 @@ app$snapshotInit("app_test")
 
 app$uploadFile(file1 = "gender_test.csv")
 app$snapshot()
-app$setInputs(dictionary = "narrow")
+app$setInputs(dictionary = "narrow_en")
 app$snapshot()
 app$snapshot()
