+++ man/narrow.Rd
@@ -1,16 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gendercodeR.R
-\docType{data}
-\name{narrow}
-\alias{narrow}
-\title{narrow}
-\description{
-<<<<<<< HEAD
-This dictionary is deprecated, please use narrow_en instead.
-}
-\details{
-A narrow English dictionary for the recode_gender function in list form
-=======
-A narrow dictionary for the recode_gender function in named vector form
->>>>>>> dev
-}
