+++ man/broad.Rd
@@ -1,16 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/gendercodeR.R
-\docType{data}
-\name{broad}
-\alias{broad}
-\title{broad}
-\description{
-<<<<<<< HEAD
-This dictionary is deprecated, please use broad_en instead.
-}
-\details{
-A broad English dictionary for the recode_gender function in list form
-=======
-A broad dictionary for the recode_gender function in named vector form
->>>>>>> dev
-}
