### 2021-08-06 016fe5f
Removing snapshot based shiny tests
