### 2021-08-10 a5bf20c
Add rationale title and make use of built-in rather than inbuilt consistent throughout. Update examples to use man/woman
