+++ R/gendercodeR.R
@@ -1,6 +1,6 @@
 #' gendercoder: A Package for Recoding Freetext Gender Data
 #'
-#' Provides built-in dictionaries and  a function \code{recode_gender}
+#' Provides dictionaries and  a function \code{recode_gender}
 #' to allow for easy automatic coding of common variations in free text
 #' responses to the question \bold{"What is your gender?"}
 #'
