+++ man/gendercoder.Rd
@@ -5,7 +5,7 @@
 \alias{gendercoder}
 \title{gendercoder: A Package for Recoding Freetext Gender Data}
 \description{
-Provides built-in dictionaries and  a function \code{recode_gender}
+Provides dictionaries and  a function \code{recode_gender}
 to allow for easy automatic coding of common variations in free text
 responses to the question \bold{"What is your gender?"}
 }
