### 2021-11-10 #435 @ekothe
Thanks everyone for your comments, suggestions, and queries. We’re very fortunate to have such expert and generous reviewers and this package will be vastly improved for your input. I apologise for the delay in putting together this response (at this point it seems old hat to blame the pandemic - but we went back into lockdown for much of this year and it was incredibly disruptive).

Responses to comments are below.

The largest change we’ve made to the package is removing the shiny app from the master branch. This feature requires more dedicated development to ensure that documentation and tests are complete to a high standard. At this stage, we feel that incorporating such changes is an unnecessary delay to having a version of the package ready for release. We deeply appreciate the comments the reviewers have provided on the shiny app and will incorporate them into the development of the app in a later release of the package.

# Response to @juliasilge
>I recommend adding a full descriptive title to the vignette. Also I would edit to always use "built-in" or "inbuilt" but not switch back and forth.
*Response:* Thank you, we agree that these changes improve clarity. We have the vignette title and descriptive text appropriately.

Corrected in a5bf20c

>I think there is a mistake in the example for the app; it says scivrs_app().
*Response:* Costs of naive borrowing. This is now corrected. However, as noted - the app has been moved to a dev-branch to allow for further development.

Corrected in c84e7a6

>I recommend adding a second, short vignette on how, when, why to add to an existing dictionary or add a new dictionary, based on the content in data/howtoaddtodictionary.R. You can link to that from the README's contributing section.

*Response:* Done.

Added in 432b45a

>I know there has been some discussion of performance previously on this thread; my opinion is that purrr::map(dictionary, tolower) and purrr::flatten_chr(recoded_list) are very reasonable in this context (purrr is pretty darn lightweight) and that the use of lists for matching is great. 👍 My opinion is that this isn't a package whose applications make it a sensible place to invest more effort in performance improvements beyond what is already here.

*Response:* The core function has been updated to use named vector dictionaries. Thanks for your advice, testing, and code @elinw and @ldecicco-USGS.

Corrected in 42e150f

>Can you add the coverage badge to your README? You don't have any tests for the app, which isn't great; the rOpenSci package guide recommends shinytest.

*Response:* Implementing shinytest via github actions has presented some unexpected challenges, partly because the app is not currently well implemented. As outlined above we have moved the shiny app to a dev branch. We have added a coverage badge for the master branch.

Coverage badge added in bfdeedc

>Something is up with the pkgdown site, as it isn't up-to-date with the package (no info on the app).
*Response:* Thank you, we’ve now implemented a github action to automate rebuilding of the pkgdown website. This should resolve this issue.

Corrected in 7bcd5cc

>Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.

*Response:* We are very pleased to acknowledge your contribution to the package and have added you using the reviewer role in the description file

Corrected in 3433bf1

>Did you consider any other names for the fill argument? It means something somewhat different in tidyr, which may be confusing to users.

*Response:* This is a great UX comment, thanks. We’ve replaced the `fill` argument with `retain_unmatched`.

Corrected in 02225cc

>My first major feedback is that this package uses the terms "male" and "female" for gender, but in my experience most practitioners are moving away from these terms for gender, instead of using them for sex. If you think about how people are more deeply understanding sex vs. gender, this package seems to be more about gender than sex (this is typically what survey questions are getting at) and better terms for that would be "man" and "woman". I strongly recommend changing this throughout the documentation and the data dictionaries.
*Response:* We have updated accordingly, as this is very much more in keeping with our intent even though it’s inconsistent with current practice. One point of tension is our ignorance of any age-agnostic gender labels, so this change has resulted in addition of additional outcome labels (i.e. male now splits into man and boy conditional on inputs).

Corrected in 2d10761f

>The second major feedback is that I find the terms "broad" and "narrow" so confusing because to me, I would have flipped them, with the broad dictionary bucketing responses into "broadly" into categories all together and the narrow dictionary keeping lots of detailed categories. Is this just me? Did you all talk about any other names for the two options that might be more "broadly" 😩 understood?
*Response:* Thank you for this feedback. When we were first working on this project was had dictionary names that corresponded to the number of categories that gender was recoded into, but that quickly because impossible as we realised the number of potential categories we needed to deal with. After much discussion among the team we’ve settled on “manylevels_en” and “fewlevels_en” to indicate the numbers of levels of the factor that are created when using each dictionary. We hope that this improves the clarity of the dictionary names and thus the usability of the package.

Corrected in 1e1fa73

>Another note is that the discussions of context in the vignette and README are great, but I would also add an acknowledgement that our understanding is changing with time and that is expected and OK, etc. You can commit to changing these mappings as our understanding grows/changes.
*Response:* Thank you for this helpful suggestion. I’ve added some text to this effect.

Corrected in

# Response to @elinw

>Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.

*Response:* We are very pleased to acknowledge your contribution to the package and have added you using the reviewer role in the description file

Corrected in 3433bf1

>I really like the concept of this package and think it will get a lot of use as the asking of open-ended gender items on surveys becomes increasingly common. This is part of a larger category of such questions. For example I recently asked about race/ethnicity in this format. I also asked the incoming cohort of first year students what major they were thinking of and that also needed coding. Further, many surveys include an “other ____” category which is often difficult to deal with. I’d imagine that this package would be extremely helpful.
Building on this comment, in reality, someone could add any dictionaries they wanted to this. That is, there is nothing inherently about gender. I think you should consider whether you rename it to something more generic and then simply say you are including the gender dictionaries as examples. I think that small change will have a huge impact on use.
*Response:* We absolutely agree that this represents one instance of a more general task, but we are concerned that making this a more general package (even retaining the recode_gender function within) we might compromise the findability and uptake of this core function. We think this is a pressing issue to address and so would rather not compromise on that core aim and instead will spin off another coding package using the same skeleton function for ethnicity and nationality (@Lingtax is particularly taken by a one line conversion from country to United Nations geoscheme / World Bank / CIA world factbook / ICANN regions).

Corrected in NA

>The other point I would make about the dictionaries you have is that they are in English. Obviously this is not something that only applies in Anglophone settings. I think that you should change the names of the dictionaries to reflect the language in some consistent pattern. I’d suggest using the ISO IDs so you can differentiate EN-AU, EN-UK and EN-US, or more importantly different versions of Spanish and Chinese.
*Response:* We have added language based suffixes according to ISO 639-1, but felt regionalisation was unnecessary for these data.

Corrected in 54a7a7a

## Shiny App
> I attempted to pull in a large csv file which meant I couldn’t see the gender column. I’m sure there is a solution to that (maybe just display the gender column once it is selected?).
*Response:*  As you can see the Shiny app was accidentally added to the main branch during review, so was half-baked in many ways. This has now been corrected as per your suggestion.

Corrected in b153f06

>I was also surprised that it didn’t support Rda or Rds file types.
*Response:* @Lingtax didn’t originally add support for R data formats as he presumed anyone using these formats would preferentially use the functions directly. The package now supports Rds, but Rda presents problems (as `Load()` returns data objects to the environment in their state as saved, so he couldn’t make shiny cooperate)

Corrected in cb1ad84

> Documentation of the use of the Shiny app would be helpful.
*Response:* The largest change we’ve made to the package is removing the shiny app from the master branch. This feature requires more dedicated development to ensure that documentation and tests are complete to a high standard. At this stage, we feel that incorporating such changes is an unnecessary delay to having a version of the package ready for release. We deeply appreciate the comments the reviewers have provided on the shiny app and will incorporate them into the development of the app in a later release of the package.

Coding
> As I mentioned I do not think a list is the best way to store the dictionaries for a number of reasons. First, based on my testing that using a named vector instead gets you substantial performance improvement. I didn’t see any major performance benefit of using a hash table, so I think the ease of use of a named vector makes it the way to go…
*Response:* The core function has been updated to use named vector dictionaries. Thanks for your advice, testing, and code @elinw and @ldecicco-USGS.

Corrected in 42e150f

> The gender_code() function should have a .data argument because you should not be required to use the pipe. (See below).
*Response:* The function has not ever required the pipe, having the vector input as the first argument making it compatible with the pipe, but still permitting users to apply the function using either $ or [] notation as shown in the updated documentation.

> One of the nice things about this package is that the code is very straightforward an cleanly written.
*Response:* Thanks! We’re pleased that the amateur process has not compromised the product.

> You have examples and documentation for the internal function which_is_na(). This means you’ll get a manual page that you don’t want. Also I think that CRAN will flag that. Also in the examples you should namespace the external functions. CRAN would also probably also like to start the function descriptions so that it has an upper case first letter.
*Response:* The documentation is removed for which_is_na().

Corrected in 9250b80

> Also try not to rely on other packages in your examples. If you do, namespace them. However think twice about using %>%. There's no reason to use it, you can change it to use a .data argument.
*Response:* Examples are now independent of other packages.

Corrected in 1dcebb8
