### 2021-11-10 #435 @juliasilge
This is looking great!

- I think “manylevels_en” and “fewlevels_en” is a big improvement and much clearer.
- Looks like you have some merge conflict leftovers here: https://ropenscilabs.github.io/gendercoder/articles/a01_rationale.html#contributing-to-this-package-1
