### 2021-11-23 #435 @ekothe
Thanks @juliasilge - I've removed the conflicted text (there was also some in the "manylevel_en.Rd"

We really appreciate the feedback on the dictionary names - even as a user of the package I'm finding them more intuitive 😄
