### 2021-11-24 #435 @juliasilge
I've got a couple of "clean up" notes from my previous review, I believe:

> You can link to that from the README's contributing section.

- Can you link to the vignette on how to add a new dictionary in the contributing section of the README?
- There are some inconsistencies in how `manylevels_en` and `fewlevels_en` are referred to in text like the README, both with and without backticks. I'd recommend choosing one and sticking with it.
- Why is the example for `recode_gender()` wrapped in `\dontrun{}`? It would be better to really run an example to demo [on the pkgdown site](https://ropenscilabs.github.io/gendercoder/reference/recode_gender.html) so folks can see the output. If this was to avoid the pipe, you already have dplyr in `Suggests` so it should be totally fine. Showing a real example is very useful and should be a high priority; having an example wrapped in `\dontrun{}` like this will be a "no go" for CRAN.
- You'll need to remove or `.Rbuildignore` that `inst.zip` file for CRAN. Or should it go somewhere else? Seems like a weird place to put it.
- Do you need readr and rlang in Suggests? I removed them from `DESCRIPTION` and ran R CMD check and everything seemed fine.
