### 2022-01-21 #435 @ldecicco-USGS
Sorry to let this linger so long. I didn't notice that both reviewers replied positively to the updates. @ekothe , there are a few minor questions/comments from @juliasilge above. I think if you take care of those we can move forward with approval!
