### 2022-03-02 #435 @ekothe
Thanks @ldecicco-USGS, @elinw and @juliasilge

I've completed the cleanup tasks suggested by @juliasilge. These have been incorporated as of commit [a6e22f9](https://github.com/ropenscilabs/gendercoder/commit/a6e22f9d4b13124b18eabab6c11dbbf61f4059ab)

> Can you link to the vignette on how to add a new dictionary in the contributing section of the README?
- Complete

> There are some inconsistencies in how manylevels_en and fewlevels_en are referred to in text like the README, both with and without backticks. I'd recommend choosing one and sticking with it.
- Complete. Now using backticks throughout

> Why is the example for recode_gender() wrapped in \dontrun{}? It would be better to really run an example to demo [on the pkgdown site](https://ropenscilabs.github.io/gendercoder/reference/recode_gender.html) so folks can see the output. If this was to avoid the pipe, you already have dplyr in Suggests so it should be totally fine. Showing a real example is very useful and should be a high priority; having an example wrapped in \dontrun{} like this will be a "no go" for CRAN.
- Complete. While dplyr is used in suggests, checks failed until I removed the pipe and specified dplyr::mutate() in this example.

> You'll need to remove or .Rbuildignore that inst.zip file for CRAN. Or should it go somewhere else? Seems like a weird place to put it.
- Complete. This has been removed.

> Do you need readr and rlang in Suggests? I removed them from DESCRIPTION and ran R CMD check and everything seemed fine.
- Complete. These have been removed.
