### 2022-03-02 #435 @juliasilge
This looks great! 🙌 If you wanted to do a tiny bit more clean-up, I would recommend removing the old `travis.yml` file. However, I approve as is.
